Updated Connector has no more tearing
Tearing issue resolved. Turns out the issue was that the image data was being split into multiple packets, so I was constantly overwriting a partial image in the buffer with a new partial image. Occasionally I would get entire images, so whole image appears to update. At other points, only the top of the image gets updated (so not quite the same as typical tearing, where the bottom image is new).